+++ core/modules/aggregator/migration_templates/d7_aggregator_item.yml
@@ -7,7 +7,7 @@ source:
 process:
   iid: iid
   fid:
-    plugin: migration
+    plugin: migration_lookup
     migration: d7_aggregator_feed
     source: fid
   title: title
