+++ core/modules/block_content/migration_templates/d6_custom_block.yml
@@ -8,7 +8,7 @@ process:
   id: bid
   info: info
   'body/format':
-    plugin: migration
+    plugin: migration_lookup
     migration: d6_filter_format
     source: format
   'body/value': body
