+++ core/modules/comment/migration_templates/d7_comment.yml
@@ -11,7 +11,7 @@ process:
   # the cid field to allow incremental migrations.
   cid: cid
   pid:
-    plugin: migration
+    plugin: migration_lookup
     migration: d7_comment
     source: pid
   entity_id: nid
