+++ core/modules/field/migration_templates/d6_field_formatter_settings.yml
@@ -15,7 +15,7 @@ process:
   # field migration.
   field_type_exists:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_field
       source:
         - field_name
@@ -29,7 +29,7 @@ process:
   entity_type: 'constants/entity_type'
   bundle:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_node_type
       source: type_name
     -
@@ -37,7 +37,7 @@ process:
       method: row
   view_mode:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_view_modes
       source:
         - view_mode
