+++ core/modules/field/migration_templates/d7_field_instance_widget_settings.yml
@@ -14,7 +14,7 @@ process:
   # field migration.
   field_type_exists:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d7_field
       source:
         - field_name
