+++ core/modules/menu_link_content/migration_templates/d7_menu_links.yml
@@ -13,7 +13,7 @@ process:
   description: description
   menu_name:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d7_menu
       source: menu_name
     -
