+++ core/modules/path/migration_templates/d6_url_alias.yml
@@ -30,7 +30,7 @@ process:
       index:
         - 1
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_node_translation
 destination:
   plugin: url_alias
