+++ core/modules/taxonomy/migration_templates/d6_taxonomy_term.yml
@@ -9,7 +9,7 @@ process:
   # the tid field to allow incremental migrations.
   tid: tid
   vid:
-    plugin: migration
+    plugin: migration_lookup
     migration: d6_taxonomy_vocabulary
     source: vid
   name: name
@@ -22,7 +22,7 @@ process:
       method: process
       source: parent
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_taxonomy_term
   parent:
     plugin: default_value
