+++ core/modules/taxonomy/migration_templates/d6_vocabulary_entity_display.yml
@@ -17,7 +17,7 @@ process:
   options: 'constants/options'
   bundle: type
   field_name:
-    plugin: migration
+    plugin: migration_lookup
     migration: d6_taxonomy_vocabulary
     source: vid
 destination:
