+++ core/modules/taxonomy/migration_templates/d6_vocabulary_entity_form_display.yml
@@ -21,7 +21,7 @@ process:
   options/weight: 'constants/options/weight'
   bundle: type
   field_name:
-    plugin: migration
+    plugin: migration_lookup
     migration: d6_taxonomy_vocabulary
     source: vid
 destination:
