+++ core/modules/taxonomy/migration_templates/d6_vocabulary_field_instance.yml
@@ -13,7 +13,7 @@ process:
   bundle: type
   field_name:
     -
-      plugin: migration
+      plugin: migration_lookup
       migration: d6_taxonomy_vocabulary
       source: vid
     -
